+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v5.6.10.0 (uncut)
+- **http4k-connect-*** - Upgrade dependencies.
+- **http4k-connect-amazon-core*** - Simpler API for retrieving the AWS credentials of a custom profile H/T @obecker
+
+### v5.6.9.0
+- **http4k-connect-amazon-apprunner*** - [New module] Client and fake
+
 ### v5.6.8.2
 - **http4k-connect-amazon-containercredentials*** - Add Kotshi adapter to Moshi instance.